--- .github/actions/compute-sdk-test/action.yml
@@ -1,12 +1,0 @@
-name: "Compute SDK Test"
-description: "Test fixture Wasm modules with Fastly C@E (From a JS Github Action)"
-inputs:
-  config:
-    description: 'Configuration JSON of the wasm module, and their expectations'
-    required: true
-  fastly_token:
-    description: '(Required if using the C@E Environment) Fastly CLI --token that allows access to deploying to C@E Services in your config'
-    required: false
-runs:
-  using: "node12"
-  main: dist/index.js
