--- .github/actions/compute-sdk-test/src/kill-process-and-wait.js
@@ -1,12 +1,0 @@
-// Kill a spawned process (e.g Viceroy), and (a)wait for it
-// to be completely killed
-const killProcessAndWait = (childProcess) => {
-  return new Promise((resolve, reject) => {
-    childProcess.on('exit', (code) => {
-      resolve('stopped with code ' + code);
-    });
-    childProcess.kill('SIGINT');
-  });
-}
-
-module.exports = killProcessAndWait;
