+++ .github/workflows/integration-tests.yml
@@ -1,16 +1,13 @@
-name: Compute TinyGo SDK Test
+name: Integration Tests
 on: [push]
 jobs:
-  sdktest:
+  integration-tests:
     runs-on: ubuntu-latest
-    concurrency: 
-      group: SdkTestJob
-      cancel-in-progress: true
     steps:
       - name: Checkout fastly/compute-sdk-go
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
       - name: Install Go
-        uses: actions/setup-go@v2
+        uses: actions/setup-go@v3
         with:
           go-version: '1.20'
       - name: Install TinyGo
@@ -24,8 +21,5 @@ jobs:
           tinygo version
           fastly version
           viceroy --version
-      - name: compute-sdk-go Integration Tests Job
-        uses: ./.github/actions/compute-sdk-test
-        id: sdktest
-        with:
-          config: './_integration_tests/sdk-test-config.json'
+      - name: Run Integration Tests
+        run: tinygo test -v -target=compute-at-edge.json ./integration_tests/...
