--- .gitignore
@@ -1,2 +1,0 @@
-_integration_tests/fixtures/kvstore/bin/main.wasm
-_integration_tests/fixtures/kvstore/pkg/kvstore.tar.gz
