--- _integration_tests/fixtures/configstore/fastly.toml
@@ -1,28 +1,0 @@
-# This file describes a Fastly Compute@Edge package. To learn more visit:
-# https://developer.fastly.com/reference/fastly-toml/
-
-authors = ["oss@fastly.com"]
-description = ""
-language = "other"
-manifest_version = 2
-name = "configstore"
-service_id = ""
-
-[local_server]
-
-  [local_server.backends]
-
-    [local_server.backends.TheOrigin]
-      url = "https://compute-sdk-test-backend.edgecompute.app/"
-
-    [local_server.backends.TheOrigin2]
-      url = "https://compute-sdk-test-backend.edgecompute.app/"
-
-    [local_server.backends.example_backend]
-      url = "https://example.org/"
-
-  [local_server.dictionaries]
-
-    [local_server.dictionaries.configstore]
-      file = "./_integration_tests/fixtures/configstore/configstore.json"
-      format = "json"
