--- _integration_tests/fixtures/configstore/main.go
@@ -1,29 +1,0 @@
-// Copyright 2022 Fastly, Inc.
-
-package main
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/fastly/compute-sdk-go/configstore"
-	"github.com/fastly/compute-sdk-go/fsthttp"
-)
-
-func main() {
-	fsthttp.ServeFunc(func(ctx context.Context, w fsthttp.ResponseWriter, _ *fsthttp.Request) {
-		d, err := configstore.Open("configstore")
-		if err != nil {
-			fsthttp.Error(w, err.Error(), fsthttp.StatusInternalServerError)
-			return
-		}
-
-		twitter, err := d.Get("twitter")
-		if err != nil {
-			fsthttp.Error(w, err.Error(), fsthttp.StatusInternalServerError)
-			return
-		}
-
-		fmt.Fprint(w, twitter)
-	})
-}
