--- _integration_tests/fixtures/hello_world/main.go
@@ -1,16 +1,0 @@
-// Copyright 2022 Fastly, Inc.
-
-package main
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/fastly/compute-sdk-go/fsthttp"
-)
-
-func main() {
-	fsthttp.ServeFunc(func(ctx context.Context, w fsthttp.ResponseWriter, r *fsthttp.Request) {
-		fmt.Fprintf(w, "Hello, TinyGo!")
-	})
-}
