--- _integration_tests/fixtures/rtlog/fastly.toml
@@ -1,22 +1,0 @@
-# This file describes a Fastly Compute@Edge package. To learn more visit:
-# https://developer.fastly.com/reference/fastly-toml/
-
-authors = ["oss@fastly.com"]
-description = ""
-language = "other"
-manifest_version = 2
-name = "rtlog"
-service_id = ""
-
-[local_server]
-
-  [local_server.backends]
-
-    [local_server.backends.TheOrigin]
-      url = "https://compute-sdk-test-backend.edgecompute.app/"
-
-    [local_server.backends.TheOrigin2]
-      url = "https://compute-sdk-test-backend.edgecompute.app/"
-
-    [local_server.backends.example_backend]
-      url = "https://example.org/"
