--- _integration_tests/fixtures/streaming_close/fastly.toml
@@ -1,25 +1,0 @@
-# This file describes a Fastly Compute@Edge package. To learn more visit:
-# https://developer.fastly.com/reference/fastly-toml/
-
-authors = ["oss@fastly.com"]
-description = ""
-language = "other"
-manifest_version = 2
-name = "streaming_close"
-service_id = ""
-
-[local_server]
-
-  [local_server.backends]
-
-    [local_server.backends.TelemetryServer]
-      url = "http://localhost:8081/"
-
-    [local_server.backends.TheOrigin]
-      url = "https://compute-sdk-test-backend.edgecompute.app/"
-
-    [local_server.backends.TheOrigin2]
-      url = "https://compute-sdk-test-backend.edgecompute.app/"
-
-    [local_server.backends.example_backend]
-      url = "https://example.org/"
